backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
transformers==4.38.0
torch
feedparser
motor
websockets
pydantic
python-dotenv
